requirements.txt
modelscope
opencv-python
networkx
thop
timm
easydict
flask
pillow
gunicorn